Bus routes drawn, bus classes refactored
I got a punch of shit working with the buses. their routes are all
drawing. we need to update the bus locatin every 5 seconds with http
calls and json parsing. For some reason right before I left the program
will not compile. some array out of bounds in a strange and distance
class file